requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
umap-learn>=0.5
matplotlib>=3.7
seaborn>=0.12
openai==2.23.0
google-genai==1.56.0
voyageai==0.3.7
cohere==5.20.5
python-dotenv>=1.0
tqdm>=4.65
pyyaml>=6.0
mord>=0.7
sentence-transformers>=3.0
peft>=0.12
torch>=2.0
